+++ pkg/files/file_test.go
@@ -1,4 +1,4 @@
-package utl
+package files
 
 import (
 	"os"
